Move time button setup into timeControls
