--- js/timeButtons.js
@@ -1,9 +1,0 @@
-import { handleTimeButton } from './timeControls.js';
-
-export function setupTimeButtons() {
-  document.addEventListener('click', (event) => {
-    const target = event.target;
-    if (!(target instanceof HTMLElement)) return;
-    handleTimeButton(target);
-  });
-}
